refactor ScheduleFormFields into own file
